Add stub conversion for tests

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsresettlementpassportapi/config/DocumentConversionConfig.kt
@@ -1,21 +1,32 @@
 package uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.config
 
+import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import software.amazon.awssdk.services.s3.S3Client
 import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.service.DocumentConversionService
+import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.service.LibreOfficeDocumentConversionService
+import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.service.StubDocumentConversionService
 import java.io.File
 import java.nio.file.Files
+
+private val logger = KotlinLogging.logger {}
 
 @Configuration
 class DocumentConversionConfig {
   @Bean
   fun documentConversionService(
-    @Value("\${temp.dir.path:-}") tempDirPath: String,
+    @Value("\${doc.conversion.use.stub:false}") useStub: Boolean,
+    @Value("\${doc.conversion.temp.dir.path:-}") tempDirPath: String,
     @Value("\${hmpps.s3.buckets.document-management.bucketName}") bucketName: String,
     s3Client: S3Client,
   ): DocumentConversionService {
+    if (useStub) {
+      logger.warn { "Using stub document conversion service" }
+      return StubDocumentConversionService(s3Client, bucketName)
+    }
+
     val tempDirFile = if (tempDirPath == "-") {
       val tempDir = Files.createTempDirectory("docs").toFile()
       tempDir.deleteOnExit()
@@ -23,6 +34,6 @@ class DocumentConversionConfig {
     } else {
       File(tempDirPath)
     }
-    return DocumentConversionService(tempDocumentDir = tempDirFile, s3Client = s3Client, bucketName = bucketName)
+    return LibreOfficeDocumentConversionService(tempDocumentDir = tempDirFile, s3Client = s3Client, bucketName = bucketName)
   }
 }
